requirements.txt
streamlit
opencv-python-headless
fer
numpy
pillow
